[fix]#25 파일이름 과 폴더명 수정
