+++ src/components/subject_index/SubjectIndexItem.tsx
@@ -1,5 +1,5 @@
 import dummy from '../../db/data.json';
-import styles from './PdfIndexItem.module.css';
+import styles from './SubjectIndexItem.module.css';
 
 export interface PdfIndexItemProps {
     circle: number;
@@ -13,7 +13,7 @@ export default function PdfIndexItem(props: PdfIndexItemProps) {
     return (
         <div>
             {data.map((data) =>
-                <div className={styles.sbj}>
+                <div className={styles.sbj} key={data.id}>
                     <h1>{data.sbj_name}</h1>
                     <button className={styles.sbj_btn}>pdf</button>
                 </div>
